fix(tests): resolve 38 TypeScript type errors in test files
- simple-pipeline.test.ts: add `as never` to mockResolvedValue/mockReturnValue
  args inside jest.mock factories; type mock refs as jest.Mock<() => Promise<unknown>>
- VideoPreview.test.tsx: cast children as React.ReactNode in mock components
  (Button, Slider, Select variants)
- export-delegation-helpers.test.ts: cast frame.data as Uint8Array for .length access
- layout-engine-integration.test.ts: use LayoutConfig/FallbackLayoutStrategy type casts
- worker-pool.test.ts: type factory return as Worker
- GridSnapStrategy.test.ts: fix estimateComplexity/detectOverlaps to match signatures
- LayoutStrategy.test.ts: add from/to to EdgeDatum missing required props
- Video.test.tsx: cast node object as PositionedNode (has x/y)
- Update TASK-0120.md: all 4 acceptance criteria [x]
- Update overview.md: Phase 23 marked complete

+++ src/workers/__tests__/export-delegation-helpers.test.ts
@@ -169,7 +169,7 @@ describe('buildFramesFromWorkerResult (private)', () => {
 
     expect(frames[0].width).toBe(w);
     expect(frames[0].height).toBe(h);
-    expect(frames[0].data.length).toBe(w * h * 4);
+    expect((frames[0].data as Uint8Array).length).toBe(w * h * 4);
     engine.dispose();
   });
